+++ CHANGELOG.md
@@ -15,3 +15,5 @@ New UI to visualize the Autosar model.
 - New API `get_property_values()` to get values of all attributes of anautosar element.
 ## [0.2.1]() - 2021-01-22
 - Small UI improvement - posibility to select search type
+- Implemented multiple type searches (by short name, by Autosar type and by regular expression on name)
+- Added menu bar - implemented them selection at runtime